+++ SmugglerCode.Blazor.UI.Sandbox/Components/Pages/Home.razor.cs
@@ -5,6 +5,7 @@ namespace SmugglerCode.Blazor.UI.Sandbox.Components.Pages
     public partial class Home : ComponentBase
     {
         private bool _show = true;
+        private string _customerName = "John Doe";
 
         private void Toggle()
         {